rename: recommendation to productRecommendation
