+++ apps/web/src/features/home/components/productRecommendation/TypeRow.tsx
